--- examples/fancy/index.js
@@ -1,7 +1,0 @@
-import React from 'react'
-import DOM from 'react-dom'
-import Pagination from './pagination'
-
-let target = document.getElementById('root')
-
-DOM.render(<Pagination />, target)
