--- examples/index.html.ejs
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Examples</title>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-    <link rel="stylesheet" href="https://code.getmdl.io/1.3.0/material.blue_grey-light_blue.min.css" />
-    <script defer src="https://code.getmdl.io/1.3.0/material.min.js"></script>
-  </head>
-
-  <body>
-    <div class="mdl-grid">
-      <div class="mdl-cell mdl-cell--12-col">
-        <h5>Examples</h5>
-        <ul class="mdl-list">
-          <% htmlWebpackPlugin.options.examples.map(function (example) { %>
-            <li class="mdl-list__item">
-              <a href="/<%- example %>/"><%- example %></a>
-            </li>
-          <% }) %>
-        </ul>
-      </div>
-    </div>
-  </body>
-</html>
