+++ package.json
@@ -2,22 +2,32 @@
   "name": "verso",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "source": "./src",
+  "main": "verso.js",
+  "module": "verso.m.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server",
-    "build": "NODE_ENV=production webpack --env.production -p --config config/webpack.lib.js"
+    "build": "make all",
+    "release": "make release",
+    "watch": "make watch",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "examples": "yarn start-storybook -p 6006 -c stories",
+    "examples:build": "cd stories && npm install && cd .. && yarn build-storybook -c stories",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   },
   "author": "Solomon Hawk <solomon.hawk@viget.com> (http://viget.com/)",
-  "license": "ISC",
+  "license": "MIT",
   "dependencies": {
-    "extract-text-webpack-plugin": "3.0.2",
+    "enzyme": "3.3.0",
+    "enzyme-adapter-react-16": "1.1.1",
     "lodash": "4.17.4",
+    "microbundle": "0.3.1",
     "prop-types": "^15.6.0",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1"
+    "raf": "3.4.0"
   },
   "devDependencies": {
+    "@storybook/cli": "3.3.10",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.0.2",
     "babel-jest": "^21.2.0",
@@ -32,15 +42,21 @@
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-prettier": "^2.3.1",
     "eslint-plugin-react": "^7.5.1",
+    "extract-text-webpack-plugin": "3.0.2",
     "html-webpack-plugin": "^2.30.1",
     "invariant": "^2.2.2",
     "jest": "^21.2.1",
     "prettier": "^1.8.2",
-    "random-words": "^0.0.1",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
     "style-loader": "^0.18.2",
     "webpack": "^3.9.1",
     "webpack-dev-server": "^2.9.5",
-    "webpack-merge": "^4.1.1"
+    "webpack-merge": "^4.1.1",
+    "@storybook/react": "^3.3.10",
+    "@storybook/addon-actions": "^3.3.10",
+    "@storybook/addon-links": "^3.3.10",
+    "@storybook/addons": "^3.3.10"
   },
   "prettier": {
     "semi": false,
