+++ stories/examples/fancy/styles.css
