+++ test/unit/index.test.js
@@ -1,0 +1,8 @@
+import Verso from 'verso'
+import { mount } from 'enzyme'
+
+describe('tests', function() {
+  it('passes', function() {
+    expect(true).toBeTruthy()
+  })
+})
